Struct pallet_oracle::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

A public part of the pallet.

pub fn add_key(origin: OriginFor<T>, public: Public) -> DispatchResult

Add public to the list of Clamor Account IDs that can validate and send unsigned transactions with signed payload

NOTE: Only the Root User of the Clamor Blockchain (i.e the local node itself) can edit this list

pub fn del_key(origin: OriginFor<T>, public: Public) -> DispatchResult

Remove a Clamor Account ID from FragKeys NOTE: Only the Root User of the Clamor Blockchain (i.e the local node itself) can call this function

pub fn store_price(
origin: OriginFor<T>,
oracle_price: OraclePrice<T::Public, T::BlockNumber>,
_signature: T::Signature
) -> DispatchResult

This function stores the original price (U256) received from the selected oracle provider.

pub fn stop_oracle(origin: OriginFor<T>, stop_it: bool) -> DispatchResult

Circuit breaker! Stop the Oracle by changing the value of OracleStop flag. true = oracle stop from the next block, false = oracle keeps running (default).

impl<T: Config> Pallet<T>

pub fn get_price() -> Result<u128, &'static str>

A helper function to allow other pallets to fetch the latest FRAG price with correct decimals.

TODO: the decimals at the moment are USDT (6 decimals). It needs to be changed when FRAG pool will be available and the second token will be known.

pub fn fetch_price_from_oracle(block_number: T::BlockNumber)

A helper function to fetch the price, sign payload and send an unsigned transaction.

pub fn get_eth_call_data(provider: &OracleProvider) -> &'static str

Get the data needed to execute eth_call to the selected oracle provider.

For Chainlink it encodes the call to latestRoundData() function from ChainLink Feed Price contract. It uses the first 4 bytes of keccak_256(latestRoundData()), padded - Use https://emn178.github.io/online-tools/keccak_256.html.

For Uniswap it encodes the call to quoteExactInputSingle function in the Quoter smart contracts that returns the amount out received for a given exact input but for a swap of a single pool: https://docs.uniswap.org/contracts/v3/reference/periphery/lens/Quoter#quoteexactinputsingle.

The pool used at this moment is ETH/USDT. (TODO: it needs to be changed when the FRAG pool will be available).

pub fn get_price_from_oracle_data(
provider: &OracleProvider,
data: Vec<u8>
) -> Result<U256, &'static str>

Fetch the latest price from the selected oracle provider.

impl<T: Config> Pallet<T>

pub fn prices() -> u128

Storage use for the latest price received from the oracle.