Type Definition pallet_oracle::pallet::EVMLinkVoting

pub type EVMLinkVoting<T: Config> = StorageMap<_GeneratedPrefixForStorageEVMLinkVoting<T>, Identity, H256, u64>;

StorageMap that maps a FRAG token locking or unlocking event to a number of votes (). The key for this map is: blake2_256(encoded(<Amount of FRAG token that was locked/unlocked, Signature written by the owner of the FRAG token on a determinstic message, Whether it was locked or unlocked, Ethereum Block Number where it was locked/unlocked>))