Crate pallet_oracle

pallet-oracle queries the ChainLink Price Feed smart contract on Ethereum to fetch the price of FRAG/USD.

It works as an Offchain Worker that will be triggered after every block, to fetch the current price.

Re-exports

pub use pallet::*;

Modules

crypto

Based on the above KeyTypeId we need to generate a pallet-specific crypto type wrappers. We can use from supported crypto kinds (sr25519, ed25519 and ecdsa) and augment the types with this pallet-specific identifier.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Enums

OracleProvider

enum that represents the price feed provider. Vec<u8> is the address of the smart contract to use.

Constants

KEY_TYPE

Defines application identifier for crypto keys of this module.

Traits

OracleContract

Traits of the Chainlink contract on the Ethereum (Goerli) network